backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Pedro <92474347+Pedroffda@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic-settings = "^2.10.1"
pydantic = "^2.11.7"
fastapi = {extras = ["standard"], version = "^0.116.1"}
sqlalchemy = "^2.0.43"
passlib = "^1.7.4"
pytz = "^2025.2"
pyjwt = "^2.10.1"
alembic = "^1.16.4"
psycopg2-binary = "^2.9.10"
uvicorn = "^0.35.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
